+++ chapter3/account_dsl_scala/account/src/main/java/ch3/account/Calculator.java
@@ -1,5 +1,5 @@
 package ch3.account;
 
 public interface Calculator {
-    float calculate(final Account account);
+    double calculate(final Account account);
 }
