+++ chapter3/account_dsl_scala/account/src/main/java/ch3/account/CalculatorImpl.java
@@ -2,8 +2,8 @@ package ch3.account;
 
 public class CalculatorImpl implements Calculator {
     @Override
-    public float calculate(final Account account) {
-        return 100f;
+    public double calculate(final Account account) {
+        return 100.00;
     }
 
 }
